fix: resolve issues with deadlocking (#13)
## Description
While attempting to migrate the prod database to the latest version of
the schema introduced in #12, I noticed that it failed due to a
deadlock. Even after temporarily disabling the WebSoc scraper and
killing all outstanding connections it would still resurface, and as it
turns out limiting the connector to one connection resolved that issue.

Things are a bit murky because the old connections may not have been
killed properly, but this should be a step in the right direction.

## Related Issue

Aims to address #5, but the jury is still out on whether this resolves
the issue.

## How Has This Been Tested?

Tested locally, migration succeeded.

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to change)

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [ ] My code involves a change to the database schema.
- [ ] My code requires a change to the documentation.