feat(search): implement resultType flag (#21)
## Description

Add the `resultType` flag that dictates what types of results are
returned by the search endpoint. Can be `course` or `instructor`.
Behavior is the same as the endpoint currently if not provided.

## How Has This Been Tested?

Tested locally with queries using the new flag.

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to change)

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [ ] My code involves a change to the database schema.
- [ ] My code requires a change to the documentation.