feat: batch endpoints for courses/instructors (#24)
## Description

Add support for fetching courses/instructors in a bulk, given a set of
IDs.

## Motivation and Context

We have been able to support batch queries for a long time but have
never done so (mostly due to my poor planning). This should remove the
need for consumers to query our API _N_ times for _N_ things, if they
already know exactly what they want.

## How Has This Been Tested?

Tested locally.

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to change)

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [ ] My code involves a change to the database schema.
- [ ] My code requires a change to the documentation.

---------

Co-authored-by: Andrew Wang <awwang3@uci.edu>