feat: deploy key manager to AWS (#36)
## Description

Deploy key manager to AWS using SST Ion. DNS is still managed through
Cloudflare.

This honestly took me way longer than it should have because turns out
SST won't verify an ACM cert using Cloudflare, so I just gave up and
manually verified one, and passed in the cert ARN.

Also created a README with remarks on the lack of support for local
development. Hopefully we'll find a way to fix that soon, but the
deployment process is just so involved.

## Motivation and Context
Cloudflare Workers do not support parts of the Node.js crypto library,
which Auth.js requires.

## How Has This Been Tested?

Go to https://dashboard.anteaterapi.com.

## Screenshots (if appropriate):

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to change)

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [ ] My code involves a change to the database schema.
- [ ] My code requires a change to the documentation.