feat(websoc): enhance WebSoc quarter parameter validation message (#76)
<!--- Provide a general summary of your changes in the Title above -->

## Description

<!--- Describe your changes in detail -->

Updated the error message for the required quarter parameter in the
WebSoc query schema. Previously, when the quarter parameter was omitted,
the error message stated: "Required". The new message is more specific,
telling the user that the quarter parameter is required.

## Related Issue

<!--- This project only accepts pull requests related to open issues -->
<!--- If suggesting a new feature or change, please discuss it in an
issue first -->
<!--- If fixing a bug, there should be an issue describing it with steps
to reproduce -->
<!--- Please link to the issue here: -->

Link to issue #72

## Motivation and Context

<!--- Why is this change required? What problem does it solve? -->

When programmers were making websoc queries without the quarter
parameter, the previous message was not helpful and unclear on what was
missing. This change makes the API more clear in showing what is
required when the quarter parameter is absent from the query.

## How Has This Been Tested?

<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran
to -->
<!--- see how your change affects other areas of the code, etc. -->

By running and hosting the API locally with Docker and pnpm, I was able
to test my changes and confirm the results I was striving for while also
confirming that there were no errors and that nothing broke in the
process. The tests I ran included accessing the API locally through
local host 8787 and testing parameters being present and absent along
with values being present and absent for those parameters.

## Screenshots (if appropriate):


![image](https://github.com/user-attachments/assets/5ec7a106-afc8-4879-8655-4fba564687fe)


![image](https://github.com/user-attachments/assets/7ddda5b2-f340-4612-9236-01c1d395a85b)



## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to change)

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [ ] My code involves a change to the database schema.
- [ ] My code requires a change to the documentation.