feat: add web column to WebSoc (#82)
<!--- Provide a general summary of your changes in the Title above -->

## Description
Modified `libwebsoc-next` to support fetching the Web column, which
links to the section's syllabus page, and is realized as the `webURL`
field in the `WebsocSection` object.

Because this column is not reflected in the XML response, which
`libwebsoc-next` parses into JSON, the module also fetches the web
version concurrently with the XML version.

As a result, the timeouts after WebSoc query invocations in the scraper
have been doubled.

## Related Issue

Closes #41.

## Motivation and Context

Unblocks icssc/AntAlmanac#1113.

## How Has This Been Tested?

Tested locally due to schema changes. Database dump available
[here](https://anteater-api-dump.s3.us-west-1.amazonaws.com/db-pr-82.sql.gz).

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to change)

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [x] My code involves a change to the database schema.
- [ ] My code requires a change to the documentation.

---------

Co-authored-by: Andrew Wang <awwang3@uci.edu>