feat(dw-scraper): better wildcard handling (#122)
## Description

This PR aims to address two key issues with our handling of wildcards.

### Department-wide wildcards

For reasons that were never documented, wildcards for an entire
department with no other qualifier (e.g. BIOSCI @) were never properly
handled. This PR implements handling for such cases by dumping all
courses in the department. This is a substantial amount of data, but no
feasible workaround exists as department-wide wildcards may cohabitate
with other complex requirements.

While implementing this feature, I noted that the department code given
by DegreeWorks is the shortened version (without whitespace). A
generated column and index thereon is added so that this data is readily
and correctly accessible by the scraper.

### Wildcard translation to RegExp

We also aim to correct our previous incorrect assumption of what the
wildcard `@` translates to. After some investigation, we present and
implement the following new assumption.

We define a **wildcard block** as one or more consecutive (i.e. not
broken by whitespace) occurrences of `@`. A wildcard block that has _n_
occurrences of `@` translates to the regular expression `\w{n,}`. It is
worth noting of course that `\w{1,}` collapses to `\w+`, which is
similar to the theory posited in #120.

So far, we have only observed occurrences of wildcard blocks with _n_
`\in` [1, 2], so the implementation reflects this fact. We argue that
the pathological cases of _n_ > 2 are unlikely to occur in practice as
there are no four-digit course numbers, and upper division courses with
significant prefix/suffix use in their numbers are almost always covered
by the _n_ = 2 case.

## Related Issue

<!--- This project only accepts pull requests related to open issues -->
<!--- If suggesting a new feature or change, please discuss it in an
issue first -->
<!--- If fixing a bug, there should be an issue describing it with steps
to reproduce -->
<!--- Please link to the issue here: -->

Closes #120; closes #123.

## Motivation and Context

fixes PP (allegedly)

## How Has This Been Tested?
Instrumented with logging locally. Degreeworks data in production is in
the process of being refreshed.

## Screenshots (if appropriate):

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [x] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to change)

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [x] My code involves a change to the database schema.
- [ ] My code requires a change to the documentation.

---------

Co-authored-by: Dante Dam <laggycomputer@yahoo.com>