fix(study-rooms): rooms with no slots (#166)
## Description

Fix issue below which failed to handle `NULL` slots when a left join
found no slots to join to.
Zod was not expecting this and throws.

The fix is to simply throw these nulls out.

## Related Issue

Fix #165.

## How Has This Been Tested?

it works now

You can simulate the error condition by deleting all study room slots
for a room.

## Types of changes

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to change)

## Checklist:

- [x] My code involves a change to the database schema.
- [ ] My code requires a change to the documentation.