feat(study-rooms): add filters, docs, URL fields (#171)
## Description

We introduce the following:
* Documentation for all previously extant study room fields.
* New parameters `dates` and `times` for `/studyRooms`.
`dates` mandates that returned slots fall on any one of the
comma-separated yyyy-mm-dd provided.
`times` mandates that returned slots fall within any one of the
comma-separated time-of-day ranges provided.
* New `url` field on returned rooms and slots.
    The link for a room is to a page for that specific room.
The link for a slot is to the same page, but with the slot selected and
focused on the button to submit the first step of the booking form.

From a technical standpoint, we have dropped the study rooms
materialized view as we can no longer effectively utilize it.

## Related Issue

Close #170.

## Motivation and Context

they said make the endpoint better

## How Has This Been Tested?

just run REST and GraphQL

## Types of changes

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to change)

## Checklist:

- [x] My code involves a change to the database schema.
- [x] My code requires a change to the documentation.

---------

Co-authored-by: Sanskar Mishra <69398106+sanskarm7@users.noreply.github.com>